+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 unreleased
 ----------
 - Use version 3.8.0 of braintree-web
+- Doesn't show Card payment option for merchants without cards enabled
 
 1.0.0-beta.4
 ------------